Move string lint rules to strings package
